Rename reducer folder to store
