+++ studio/src/pages/RequestorPage/RequestPanel/RequestPanel.tsx
@@ -8,7 +8,7 @@ import { FormDataForm } from "../FormDataForm";
 import { KeyValueForm } from "../KeyValueForm";
 import { CustomTabTrigger, CustomTabsContent, CustomTabsList } from "../Tabs";
 import type { AiTestingPersona } from "../ai";
-import type { RequestorBody, RequestsPanelTab } from "../reducer";
+import type { RequestorBody, RequestsPanelTab } from "../store";
 import type { WebSocketState } from "../useMakeWebsocketRequest";
 import { AiDropDownMenu } from "./AiDropDownMenu";
 import { AIGeneratedInputsBanner } from "./AiGeneratedInputsBanner";
@@ -18,7 +18,7 @@ import { PathParamForm } from "./PathParamForm";
 import "./styles.css";
 import { CodeMirrorJsonEditor } from "@/components/Timeline";
 import { useShallow } from "zustand/react/shallow";
-import { useRequestorStore } from "../reducer";
+import { useRequestorStore } from "../store";
 
 type RequestPanelProps = {
   aiEnabled: boolean;
