+++ studio/src/pages/RequestorPage/ai/AiTestGenerationPanel.tsx
@@ -11,8 +11,8 @@ import {
 import { useMemo, useState } from "react";
 import { CustomTabTrigger, CustomTabsContent, CustomTabsList } from "../Tabs";
 import type { Requestornator } from "../queries";
-import { useActiveRoute, useServiceBaseUrl } from "../reducer";
 import { findMatchedRoute } from "../routes";
+import { useActiveRoute, useServiceBaseUrl } from "../store";
 import { ContextEntry } from "./AiTestGenerationDrawer";
 import { usePrompt } from "./ai-test-generation";
