+++ studio/src/pages/RequestorPage/ai/generate-request-data.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 import type { Requestornator } from "../queries";
-import type { RequestBodyType } from "../reducer";
+import type { RequestBodyType } from "../store";
 import type { ProbedRoute } from "../types";
 import { simplifyHistoryEntry } from "./utils";
