+++ studio/src/pages/RequestorPage/useRequestorHistory.ts
@@ -4,8 +4,8 @@ import { useMemo } from "react";
 import { useShallow } from "zustand/react/shallow";
 import { createKeyValueParameters } from "./KeyValueForm";
 import { type Requestornator, useFetchRequestorRequests } from "./queries";
-import { useRequestorStore } from "./reducer";
 import { findMatchedRoute } from "./routes";
+import { useRequestorStore } from "./store";
 import {
   // type RequestMethodInputValue,
   isRequestMethod,
