+++ studio/src/pages/RequestorPage/useRequestorSubmitHandler.ts
@@ -3,9 +3,9 @@ import { useHandler } from "@fiberplane/hooks";
 import { useShallow } from "zustand/react/shallow";
 import type { KeyValueParameter } from "./KeyValueForm";
 import type { MakeProxiedRequestQueryFn } from "./queries";
-import type { RequestorBody } from "./reducer";
-import { useRequestorStore } from "./reducer";
-import { useServiceBaseUrl } from "./reducer/useServiceBaseUrl";
+import type { RequestorBody } from "./store";
+import { useRequestorStore } from "./store";
+import { useServiceBaseUrl } from "./store/useServiceBaseUrl";
 import { isWsRequest } from "./types";
 
 export function useRequestorSubmitHandler({
